+++ internal/transport/http/handler/handler.go
@@ -23,7 +23,7 @@ type Handler struct {
 type Handlerer interface {
 	SetUpdates(metrics []dto.Metric) ([]models.Metric, error)
 	SetMetric(metric dto.Metric) (*models.Metric, error)
-	GetMetricValue(name string, typeStr string) (*int64, error)
+	GetMetricValue(name string, typeStr string) (*float64, error)
 	GetHTML(w http.ResponseWriter) error
 }
